requirements.txt
fastapi
uvicorn
torch
transformers
numpy
pandas
scikit-learn
pillow
tqdm